openhab: Remove wrapper functions.
Not necessary anymore, as the legacy type mapping is used again.

+++ openhab/StandardFlashHost.java
@@ -10,9 +10,9 @@ import org.eclipse.smarthome.core.thing.binding.firmware.Firmware;
 import org.eclipse.smarthome.core.thing.binding.firmware.ProgressCallback;
 import org.eclipse.smarthome.core.thing.binding.firmware.ProgressStep;
 public interface StandardFlashHost {
-    public abstract void writeBrickletPlugin(char port, int chunkOffset, int[] chunk) throws TinkerforgeException;
+    public abstract void writeBrickletPlugin(char port, short chunkOffset, short[] chunk) throws TinkerforgeException;
 
-    public abstract int[] readBrickletPlugin(char port, int chunkOffset) throws TinkerforgeException;
+    public abstract short[] readBrickletPlugin(char port, short chunkOffset) throws TinkerforgeException;
 
     public abstract void reset() throws TinkerforgeException;
 
@@ -25,20 +25,20 @@ public interface StandardFlashHost {
 
         progressCallback.next();
 
-        List<int[]> plugin_chunks = new ArrayList<>();
+        List<short[]> plugin_chunks = new ArrayList<>();
         int offset = 0;
 
         int PLUGIN_CHUNK_SIZE = 32;  //IPConnection.PLUGIN_CHUNK_SIZE
         while(offset < plugin.length) {
-            int[] chunk = new int[PLUGIN_CHUNK_SIZE];
+            short[] chunk = new short[PLUGIN_CHUNK_SIZE];
             for(int i = 0; i < Math.min(plugin.length - offset, PLUGIN_CHUNK_SIZE); ++i)
                 chunk[i] = plugin[offset + i];
             plugin_chunks.add(chunk);
             offset += PLUGIN_CHUNK_SIZE;
         }
 
-        int position = 0;
-        for(int[] chunk : plugin_chunks) {
+        short position = 0;
+        for(short[] chunk : plugin_chunks) {
             try {
                 this.writeBrickletPlugin(port, position, chunk);
             } catch (TinkerforgeException e) {
@@ -53,8 +53,8 @@ public interface StandardFlashHost {
         progressCallback.next();
 
         position = 0;
-        for(int[] chunk : plugin_chunks) {
-            int[] actual;
+        for(short[] chunk : plugin_chunks) {
+            short[] actual;
             try {
                 actual = this.readBrickletPlugin(port, position);
             } catch (TinkerforgeException e) {
